mismaHora? terminado, faltan testeos

+++ db/migrate/20180709213509_add_duracion_to_trips.rb
@@ -1,5 +1,5 @@
 class AddDuracionToTrips < ActiveRecord::Migration[5.0]
   def change
-  	  add_column :trips, :duracion, :float
+  	  add_column :trips, :duracion, :integer
   end
 end
